% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{append_string}
\alias{append_string}
\title{Append string to each element of character vector}
\usage{
append_string(char_vector, string_to_append)
}
\arguments{
\item{char_vector}{A character vector}

\item{string_to_prepend}{String to prepend}
}
\value{
Character vector with string appended to each element
}
\description{
Append string to each element of character vector
}

