% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{cast_and_quote}
\alias{cast_and_quote}
\title{Wrap single quotes around values for character, Date, factor, ordered
factor fields and POSIXct date or date/time vectors}
\usage{
cast_and_quote(x, explicit_type_cast = NULL)
}
\arguments{
\item{x}{A vector}

\item{explicit_type_cast}{A type cast specification (specify :: if required)}
}
\value{
The same vector with single quotes around values if needed e.g. 
to use in a SQL statement. Type casts for most standard types will be automatically
performed. If a custom type cast is desired, it can be specified
}
\description{
Useful for creating SQL strings.
}
\examples{
cast_and_quote(CO2$Treatment, "::varchar") # Explicit type cast
}

