% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{cast_and_quote_df}
\alias{cast_and_quote_df}
\title{Wrap single quotes around data frame columns and cast appropriately}
\usage{
cast_and_quote_df(df, explicit_type_cast_list = NULL)
}
\arguments{
\item{x}{A vector}

\item{explicit_type_cast}{A type cast specification (specify :: if required)}
}
\value{
The same vector with single quotes around values if needed e.g. 
to use in a SQL statement. Type casts for most standard types will be automatically
performed. If a custom type cast is desired, it can be specified
}
\description{
Useful for creating SQL strings.
}
\examples{
cast_and_quote(CO2$Treatment, "::varchar") # Explicit type cast
}

