% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{create_matching_where}
\alias{create_matching_where}
\title{Create a SQL WHERE clause defining table rows matching a given data frame.}
\usage{
create_matching_where(df, pk_fields, explicit_type_cast_list = NULL)
}
\arguments{
\item{df}{A data frame containing primary keys with which to query the db.}

\item{pk_fields}{A character vector naming the fields which comprise a
primary key for selecting matching records in a SQL database table.}
}
\value{
A string with a SQL WHERE clause describing matching records.
}
\description{
Creates a WHERE clause which can be added to a SQL statement such as
SELECT or UPDATE in order to define rows in a database as matching those
in a given data frame.  Useful for creating custom UPSERT operations. Will
wrap single quotes around values for character, Date, factor and ordered
factor fields.
}
\examples{
create_matching_where(CO2, c("Plant", "Type", "Treatment", "conc"))
}

