% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{dbWriteTable2_fast_concurrent}
\alias{dbWriteTable2_fast_concurrent}
\title{Write table to postgres database with auto-incrementing primary key}
\usage{
dbWriteTable2_fast_concurrent(con, table.name, df, fill.null = TRUE,
  add.id = TRUE, row.names = FALSE, pg.update.seq = TRUE, ...)
}
\arguments{
\item{con}{A DBI connection object}

\item{df}{Dataframe to write}

\item{fill.null}{Should NAs be filled?}

\item{add.id}{Should ids be added?}

\item{row.names}{Should row names be written?}

\item{pg.update.seq}{Should the serial sequencer be updated for postgres tables}
}
\value{
Return value indicating succss or failure
}
\description{
Write table to postgres database with auto-incrementing primary key
}

