% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{get_matching_records}
\alias{get_matching_records}
\title{Return sets of primary keys from a db matching primary keys in a data frame.}
\usage{
get_matching_records(con, df, table, pk_fields,
  explicit_type_cast_list = NULL)
}
\arguments{
\item{con}{A connection object.}

\item{df}{A data frame.}

\item{table}{A string naming the target table in the database.}

\item{pk_fields}{A character vector naming the fields which comprise a
primary key for selecting matching records in a SQL database table.}
}
\value{
A data frame of matching primary key fields from the database with row counts
}
\description{
Return sets of primary keys from a db matching primary keys in a data frame.
}

