% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{key_value_pairs_for_sql}
\alias{key_value_pairs_for_sql}
\title{Create a set of key-value pair strings for use in constructing SQL queries.}
\usage{
key_value_pairs_for_sql(df, fields, explicit_type_cast_list = NULL)
}
\arguments{
\item{df}{A data frame}

\item{fields}{A string identifying which field in the data frame to be used}
}
\value{
A character vector of key-value pairs
}
\description{
Intended for use in creating WHERE statments as well as UPSERT operations.
}
\examples{
key_value_pairs(CO2, "Treatment")
key_value_pairs(CO2, "uptake")
}

