% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{matching_records_builder}
\alias{matching_records_builder}
\title{Match records in a data frame with db records by primary keys batch by batch.}
\usage{
matching_records_builder(con, df, table, pk_fields,
  explicit_type_cast_list = NULL, batch_size = 1000, verbose = FALSE)
}
\arguments{
\item{con}{A connection object.}

\item{batch_size}{The number of records to match at once.}

\item{import}{A data frame.}

\item{...}{Additional arguments to pass to get_matching_records.}
}
\value{
A data frame of row identifier keys found in both the df and the database with
the row counts for each combination in the db
}
\description{
Gets around the stack depth limit encountered when passing a very long WHERE
clause to SQL. Useful for planning an upsert operation (esp. separating 
updates from inserts).
}

