% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{prepend_string}
\alias{prepend_string}
\title{Prepend string to each element of character vector}
\usage{
prepend_string(char_vector, string_to_prepend)
}
\arguments{
\item{char_vector}{A character vector}

\item{string_to_prepend}{String to prepend}
}
\value{
Character vector with string prepended to each element
}
\description{
Prepend string to each element of character vector
}

