% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{quote_string}
\alias{quote_string}
\title{Wrap single quote around string}
\usage{
quote_string(char_vector)
}
\arguments{
\item{char_vector}{A character vector}
}
\value{
Character vector with each element wrapped in single quotes
}
\description{
Wrap single quote around string
}

