% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{write_update_statement}
\alias{write_update_statement}
\title{Create a SQL UPDATE statement from a data frame.}
\usage{
write_update_statement(df, table, pk_fields, set_fields,
  explicit_type_cast_list = NULL)
}
\arguments{
\item{df}{A data frame containing the primary keys and values for the update.}

\item{table}{A string naming the target table in the database.}

\item{pk_fields}{A character vector naming the fields which comprise a
primary key for selecting matching records in a SQL database table.}

\item{set_fields}{A character vector naming the fields to be updated.}

\item{...}{Additional arguments esp. postgres_types}
}
\value{
A string comprising a SQL update statement.
}
\description{
Useful for creating custom UPSERT operations. Will wrap single quotes around 
values for character, Date, factor and ordered factor fields.
Implements answer to http://stackoverflow.com/questions/18797608.
Currently converts all fields to text in the where clause to obviate need
to specify types for Postgres.
}
\examples{
write_update_statement(CO2, "co2", c("Plant", "Type"), c("conc", "uptake"))
write_update_statement(CO2, "co2", c("Type"), c("uptake"), c("TEXT", NA))
}

