% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upseRt.R
\name{write_values_tuples}
\alias{write_values_tuples}
\title{Create a set of tuples from rows in a data frame.}
\usage{
write_values_tuples(df, explicit_type_cast_list = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Additional arguments esp. postgres_types}
}
\value{
A single string with a comma-separated list of values tuples.
}
\description{
Intended for use in creating a SQL update statement. Set postgres_types only
if needed to match data type esp. for dates, timestamps.
See format at http://www.postgresql.org/docs/current/static/sql-values.html.
See also http://www.postgresql.org/docs/9.4/static/datatype.html for types.
}
\examples{
write_values_tuples(head(CO2, 10), c(NA, NA, NA, "text", NA))
}

